Changed some variable types - João Pedro

+++ src/process.c
@@ -2,9 +2,9 @@
 
 Img *greyscale(Img *img)
 {
-	unsigned int toned;
-	for (int i = 0; i < img->height; ++i) {
-		for (int j = 0; j < img->width; ++j) {
+	ushort toned;
+	for (ushort i = 0; i < img->height; ++i) {
+		for (ushort j = 0; j < img->width; ++j) {
 			toned = (img->pixels[i][j].r*0.3 + img->pixels[i][j].g*0.59 + img->pixels[i][j].b*0.11);
 			img->pixels[i][j].r = toned;
 			img->pixels[i][j].g = toned;
@@ -17,7 +17,7 @@ Img *greyscale(Img *img)
 	return img;
 }
 
-Img *gaussianFilter(Img *originalImg, short limit)
+Img *gaussianFilter(Img *originalImg, uchar limit)
 {
 	if (limit <= 0) {
 		fprintf(stderr, "Can't execute %i times the Gaussian filter, please give a valid number (greater than 0).\n", limit);
@@ -26,23 +26,23 @@ Img *gaussianFilter(Img *originalImg, short limit)
 
 	Img *gaussImg = createImg(originalImg->height, originalImg->width);
 
-	int gauss[7][7] = {{1, 1, 2, 2, 2, 1, 1}, {1, 3, 4, 5, 4, 3, 1}, {2, 4, 7, 8, 7, 4, 2}, {2, 5, 8, 10, 8, 5, 2}, {2, 4, 7, 8, 7, 4, 2}, {1, 3, 4, 5, 4, 3, 1}, {1, 1, 2, 2, 2, 1, 1}};
-	double filter;
-	double normalizationFactor = 170;
-	static unsigned short count = 0;
+	uchar gauss[7][7] = {{1, 1, 2, 2, 2, 1, 1}, {1, 3, 4, 5, 4, 3, 1}, {2, 4, 7, 8, 7, 4, 2}, {2, 5, 8, 10, 8, 5, 2}, {2, 4, 7, 8, 7, 4, 2}, {1, 3, 4, 5, 4, 3, 1}, {1, 1, 2, 2, 2, 1, 1}};
+	float filter;
+	float normalizationFactor = 170;
+	static uchar count = 0;
 
 	// uncomment these two lines if you want to use a 5x5 Gaussian kernel
 	// int gauss[5][5] = {{2, 4, 5, 4, 2}, {4, 9, 12, 9, 4}, {5, 12, 12, 12, 5}, {4, 9, 12, 9, 4}, {2, 4, 5, 4, 2}};
 	// double normalizationFactor = 159;
 
-	for (int i = 0; i < gaussImg->height; i++) {
-		for (int j = 0; j < gaussImg->width; j++) {
+	for (ushort i = 0; i < gaussImg->height; i++) {
+		for (ushort j = 0; j < gaussImg->width; j++) {
 			// this will get each pixel of img->**pixels
 			// each pixel will be the "15" on the matrix (gauss[2][2])
 
 			filter = 0;
-			for (int k = 0; k < 7; ++k) {
-				for (int l = 0; l < 7; ++l) {
+			for (uchar k = 0; k < 7; ++k) {
+				for (uchar l = 0; l < 7; ++l) {
 					if ( ((i-3+k) >= 0 && (j-3+l) >= 0) && (i+3 < gaussImg->height && j+3 < gaussImg->width) ) {
 						filter += originalImg->pixels[i-3+k][j-3+l].r * gauss[k][l]/normalizationFactor;
 						// printf("filter executed\n");
@@ -79,7 +79,7 @@ Img *gaussianFilter(Img *originalImg, short limit)
 	return gaussImg;
 }
 
-Img *sobelFilter(Img *originalImg, short limit)
+Img *sobelFilter(Img *originalImg, uchar limit)
 {
 	if (limit <= 0) {
 		fprintf(stderr, "Can't execute %i times the Sobel filter, please give a valid number (greater than 0).\n", limit);
@@ -92,10 +92,10 @@ Img *sobelFilter(Img *originalImg, short limit)
 	int sobel_y[3][3] = {{1, 2, 1}, {0, 0, 0}, {-1, -2, -1}};
 	int filter_x;
 	int filter_y;
-	static unsigned short count = 0;
+	static uchar count = 0;
 
-	for (int i = 0; i < sobelImg->height; ++i) {
-		for (int j = 0; j < sobelImg->width; ++j) {
+	for (ushort i = 0; i < sobelImg->height; ++i) {
+		for (ushort j = 0; j < sobelImg->width; ++j) {
 			// this will get each pixel of img->**pixels
 			// each pixel will be on the "0" on either matrix (sobel_x[1][1] and
 			// sobel_y[1][1]
@@ -103,8 +103,8 @@ Img *sobelFilter(Img *originalImg, short limit)
 			// sobel_x and sobel_y covolution
 			filter_x = 0;
 			filter_y = 0;
-			for (int k = 0; k < 3; ++k) {
-				for (int l = 0; l < 3; ++l) {
+			for (uchar k = 0; k < 3; ++k) {
+				for (uchar l = 0; l < 3; ++l) {
 					// this tests if the pixel is on a border
 					if ( (i == 0 || j == 0) || (i == sobelImg->height - 1 || j == sobelImg->width - 1) ) {
 						// if it is, then its value is 0
